requirements.txt
easydict>=1.9,<2
pytorch>=1.1,<2
opencv>=4.1.0<5.0
numpy>=1.16,<2.0
torchvision>=0.3